+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Bypass Tokio `RecvError`. We control the `Sender`, so it should never be
   dropped too early.
 - Remove `prelude` module.
+- Seal `AsyncThreadPool` trait.
 
 ## 1.0.0